% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_u2.R
\name{adj_u2}
\alias{adj_u2}
\title{Adjust wind speed data}
\usage{
adj_u2(u, uz)
}
\arguments{
\item{u}{measured wind speed at uz above ground surface [m/s]}

\item{uz}{height of windspeed measurement above ground surface [m]}
}
\description{
Adjust wind speed data to 2 meter height.
}
\note{
eq. 47 of reference
}
\examples{
adj_u2(3.2, uz = 10)
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56. FAO, Rome, 300(9).
}
