% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo.R
\name{e0}
\alias{e0}
\title{Saturation Vapor Pressure}
\usage{
e0(Temp)
}
\arguments{
\item{Temp}{Temperature [degreeC]}
}
\value{
saturation vapour pressure at air temperature [kPa/degreeC])
}
\description{
Mean saturation vapour pressure [kPa] for short time intervals less than a day. Calculated with air temperature.
}
\note{
eq. 11 of the reference

for day, week, decade or month, the mean saturation vapour pressure should be computed with \code{\link{satVP}}
}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56. FAO, Rome, 300(9).
}
\seealso{
\code{\link{satVP}}, \code{\link{VP}}
}
