% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdate.r
\name{prep.date}
\alias{prep.date}
\title{Prepare date}
\usage{
prep.date(x)
}
\arguments{
\item{x}{date-time object or Day of the year}
}
\value{
Day of the year
}
\description{
Checks if x is date-time object or day of the year (doy) and returns doy.
}
